Due to permission errors and to security, we will work on /mnt/diskS/ instead of gpfs.
This is because when building singularity images we require of `sudo` permissions. Even though we could use `--fakeroot` this gets in conflict with the exporting of the `tmp` and `cache` dirs to other non-standard directories. Therefore, we are now mounting everything, tmp and cache included, onto /mnt/diskS/ which in theory we should not ran out of space if done properly.

Also this is nice because we will have a centralized cache for all iamges, speeding up a lot on some recurrent builds